@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Data;
 using Hellang.Middleware.ProblemDetails;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
@@ -10,48 +9,35 @@ namespace ProblemDetailsDemo.Api.ProblemDetailsConfig
 {
     public class ProblemDetailsOptionsCustomSetup : IConfigureOptions<ProblemDetailsOptions>
     {
-        public ProblemDetailsOptionsCustomSetup(IHostingEnvironment environment,
-            IHttpContextAccessor httpContextAccessor, IOptions<ApiBehaviorOptions> apiOptions)
+        public ProblemDetailsOptionsCustomSetup(IOptions<ApiBehaviorOptions> apiOptions)
         {
-            Environment = environment;
-            HttpContextAccessor = httpContextAccessor;
-            ApiOptions = apiOptions.Value;
+          ApiOptions = apiOptions.Value;
         }
 
-        private IHostingEnvironment Environment { get; }
-        private IHttpContextAccessor HttpContextAccessor { get; }
         private ApiBehaviorOptions ApiOptions { get; }
 
         public void Configure(ProblemDetailsOptions options)
         {
-            options.IncludeExceptionDetails = ctx => Environment.IsDevelopment();
-
             options.MapStatusCode = MapStatusCode;
-
-            options.OnBeforeWriteDetails = (ctx, details) =>
-            {
-                // keep consistent with asp.net core 2.2 conventions that adds a tracing value
-                ProblemDetailsHelper.SetTraceId(details, HttpContextAccessor.HttpContext);
-            };
 
             // This will map DBConcurrencyException to the 409 Conflict status code.
             options.Map<DBConcurrencyException>(ex =>
-                new ExceptionProblemDetails(ex, StatusCodes.Status409Conflict));
+                new StatusCodeProblemDetails(StatusCodes.Status409Conflict));
 
             // This will map NotImplementedException to the 501 Not Implemented status code.
             options.Map<NotImplementedException>(ex =>
-                new ExceptionProblemDetails(ex, StatusCodes.Status501NotImplemented));
+                new StatusCodeProblemDetails(StatusCodes.Status501NotImplemented));
         }
 
-        private ProblemDetails MapStatusCode(HttpContext context, int statusCode)
+        private ProblemDetails MapStatusCode(HttpContext context)
         {
             if (!ApiOptions.SuppressMapClientErrors &&
-                ApiOptions.ClientErrorMapping.TryGetValue(statusCode, out var errorData))
+                ApiOptions.ClientErrorMapping.TryGetValue(context.Response.StatusCode, out var errorData))
             {
                 // prefer the built-in mapping in asp.net core
                 return new ProblemDetails
                 {
-                    Status = statusCode,
+                    Status = context.Response.StatusCode,
                     Title = errorData.Title,
                     Type = errorData.Link
                 };
@@ -59,7 +45,7 @@ namespace ProblemDetailsDemo.Api.ProblemDetailsConfig
             else
             {
                 // use Hellang.Middleware.ProblemDetails mapping
-                return new StatusCodeProblemDetails(statusCode);
+                return StatusCodeProblemDetails.Create(context.Response.StatusCode);
             }
         }
     }
