@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 namespace ProblemDetailsDemo.Api
 {
@@ -10,11 +10,10 @@ namespace ProblemDetailsDemo.Api
             CreateWebHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-        {
-            return WebHost.CreateDefaultBuilder(args)
-//                .UseEnvironment(EnvironmentName.Production) // Uncomment to remove exception details from responses.
-                .UseStartup<Startup>();
-        }
+        public static IHostBuilder CreateWebHostBuilder(string[] args) => Host.CreateDefaultBuilder(args)
+          .ConfigureWebHostDefaults(webBuilder => webBuilder
+            .UseStartup<Startup>()
+            // .UseEnvironment(Environments.Production) // Uncomment to remove exception details from responses.
+          );
     }
 }
